네이밍수정 및 파일위치 변경
