+++ LionHeart-iOS/LionHeart-iOS/Scenes/Splash/SplashViewController.swift
@@ -11,6 +11,19 @@ import UIKit
 import SnapKit
 import Lottie
 
+protocol UserInProtocol {
+    func reissueToken(token: Token) async throws -> Token?
+    func login(type: LoginType, kakaoToken: String) async throws
+    func signUp(type: LoginType, onboardingModel: UserOnboardingModel) async throws
+}
+
+protocol UserOutProtocol {
+    func resignUser() async throws
+    func logout(token: UserDefaultToken) async throws
+}
+
+protocol AuthServiceProtocol: UserInProtocol, UserOutProtocol {}
+
 final class SplashViewController: UIViewController {
 
     // MARK: - UI Components
